+++ lib/parsedate.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: parsedate.c,v 1.15 2005-03-31 07:02:03 bagder Exp $
+ * $Id: parsedate.c,v 1.16 2005-04-26 13:08:49 bagder Exp $
  ***************************************************************************/
 /*
   A brief summary of the date string formats this parser groks:
@@ -398,7 +398,7 @@ static time_t Curl_parsedate(const char *date)
 
     /* Add the time zone diff (between the given timezone and GMT) and the
        diff between the local time zone and GMT. */
-    delta = (tzoff!=-1?tzoff:0) + (t - t2);
+    delta = (long)((tzoff!=-1?tzoff:0) + (t - t2));
 
     if((delta>0) && (t + delta < t))
       return -1; /* time_t overflow */
